requirements.txt
langchain-groq>=0.2.0
langgraph>=0.2.0
langchain>=0.3.0
streamlit>=1.40.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
